Ignore overrides of the crates.io registry from the cargo config
If the crates.io registry is overridden currently, we'll end up trying
to look up information from the replacement registry rather than the
original crates.io registry. As the logic in cargo-vet is dependent on
operating against the real crates.io registry, this forces the use of
the true registry for internal checks against the index.

Fixes #475