Switch resolver to using guppy
This switches the backend used for building and managing the dependency
graph from a custom-built backend based on the cargo metadata output to
instead be based on the guppy crate.

Guppy ingests cargo metadata output and combines that output with
re-implementations of algorithms from Cargo in order to allow simulating
cargo builds more accurately, including details like platform-specific
configs and feature defines.

This migration does not add any additional feature support for
specifying platforms to perform audits for and/or for limiting
feature/platform support when performing audits, but will provide the
framework to make these changes possible.

Due to the fairly fundamental nature of the dependency graph, this
change introduces a number of BREAKING CHANGES, which have been
documented in the CHANGELOG.md.